+++ packages/synapse-interface/slices/portfolio/actions.ts
@@ -18,6 +18,10 @@ export enum FetchState {
 export const setActiveTab = createAction<PortfolioTabs>(
   'portfolio/setActiveTab'
 )
+export const typeSearchInput = createAction<{ searchInput: string }>(
+  'portfolio/typeSearchInput'
+)
+
 export const updateSingleTokenAllowance = createAction<{
   chainId: number
   allowance: bigint
@@ -27,3 +31,4 @@ export const updateSingleTokenAllowance = createAction<{
 export const resetPortfolioState = createAction<void>(
   'portfolio/resetPortfolioState'
 )
+export const resetSearchState = createAction<void>('portfolio/resetSearchState')