+++ packages/synapse-interface/slices/swap/hooks.ts
@@ -1,0 +1,6 @@
+import { RootState } from '@/store/store'
+import { useAppSelector } from '@/store/hooks'
+
+export const useSwapState = (): RootState['swap'] => {
+  return useAppSelector((state) => state.swap)
+}